+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added missing menu actions to the last user message
 - correctly handling copy actions on MacOS and mobile devices
+- correctly handling export to markdown in browsers
 
 ## [0.47.0]